pyproject.toml
[project]
name = "enterprise-pipeline"
version = "0.1.0"
description = "Enterprise AI Pipeline — RAG with safety, observability, and compliance"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "unstructured[all-docs]>=0.16,<1",
    "python-magic>=0.4,<1",
    "qdrant-client>=1.13,<2",
    "openai>=1.60,<2",
    "tiktoken>=0.8,<1",
    "numpy>=2.2,<3",
    "rank-bm25>=0.2,<1",
    "spacy>=3.8,<4",
    "cohere>=5.13,<6",
    "langfuse>=2.57,<3",
    "structlog>=24.4,<25",
    "pyyaml>=6.0,<7",
    "deepmerge>=2.0,<3",
    "httpx>=0.28,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6.0,<7",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
    "respx>=0.22,<1",
    "types-PyYAML>=6.0,<7",
]
eval = [
    "deepeval>=2.5,<3",
    "ragas>=0.2,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "eval: marks tests as evaluation tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "langfuse",
    "langfuse.*",
    "rank_bm25",
    "rank_bm25.*",
    "cohere",
    "cohere.*",
    "unstructured",
    "unstructured.*",
]
ignore_missing_imports = true
